Update (#43)
 We apologise: this update brings Decoder API changes =(

Other changes:
 * fix spelling
 * fix some clang-tidy warnings
 * Java: move `org.brunsli` -> `dev.brunsli`
 * pull some internal API to `(enc|dec)/state.h`

+++ c/common/constants.h
@@ -69,8 +69,8 @@ static const uint8_t kBrunsliSignature[] = {
 };
 static const size_t kBrunsliSignatureSize = sizeof(kBrunsliSignature);
 
-static const int kMaxApp0Densities = 8;
-static const int kApp0Densities[kMaxApp0Densities] = {
+static const size_t kMaxApp0Densities = 8;
+static const uint16_t kApp0Densities[kMaxApp0Densities] = {
   1, 72, 96, 100, 150, 180, 240, 300
 };
